fixed: Separate modals use Sucess modals fix reducer
