+++ src/components/hanseithon/current/ht-current.component.tsx
@@ -50,6 +50,27 @@ const CategoryBox = styled.div<{ active: boolean }>`
   }
 `;
 
+const CurrentCategoryBox = styled.div<{ active: boolean }>`
+  width: 25%;
+
+  cursor: pointer;
+
+  font-family: 'Opne Sans';
+  font-size: 1.375rem;
+  font-weight: bold;
+  color: ${props => (props.active ? '#ffffff' : '#000000')};
+
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  background-color: ${props => (props.active ? '#000000' : '#ffffff')};
+
+  @media ${Device.mobileL} {
+    font-size: 1.12rem;
+  }
+`;
+
 const ListWrapper = styled.div`
   width: 100%;
 
@@ -76,15 +97,21 @@ const ListWrapper = styled.div`
   }
 `;
 
-const HTCurrentComponent: React.FC<HTCurrentMethod & HTCurrentProps> = ({
+type MatchType = 'match-l' | 'match-g';
+
+const HTCurrentComponent: React.FC<
+  HTCurrentMethod & HTCurrentProps & { isModal?: boolean }
+> = ({
   teams,
   getTeam,
   getTeamStatus,
   accessToken,
   modal,
   setTeamPk,
+  isModal,
 }) => {
-  const [category, setCategory] = useState<CategoryType>('l');
+  const [category, setCategory] = useState<CategoryType | MatchType>('l');
+
   const TeamsList =
     getTeamStatus === 'success'
       ? teams.map((item, i) => {
@@ -103,21 +130,82 @@ const HTCurrentComponent: React.FC<HTCurrentMethod & HTCurrentProps> = ({
         })
       : [];
 
+  const MatchList =
+    getTeamStatus === 'success'
+      ? teams.map((item, i) => {
+          return (
+            <HTCurrentItem
+              key={i}
+              name={item.name}
+              leaderName={item.leader_name}
+              teamMember={item.teamMember}
+              team_pk={item.pk}
+              category={category}
+              modal={modal}
+              setTeamPk={setTeamPk}
+            />
+          );
+        })
+      : [];
+
   useEffect(() => {
-    getTeam({ accessToken, category });
+    if (category === 'l' || category === 'g') {
+      getTeam({ accessToken, category });
+    } else if (category === 'match-i' || category === 'match-g') {
+      // 매칭 요청
+    }
   }, [category]);
 
   return (
     <Wrapper>
       <CategoryWrapper>
-        <CategoryBox active={category === 'l'} onClick={() => setCategory('l')}>
-          생활 부문
-        </CategoryBox>
-        <CategoryBox active={category === 'g'} onClick={() => setCategory('g')}>
-          게임 부문
-        </CategoryBox>
+        {isModal ? (
+          <>
+            <CategoryBox
+              active={category === 'l'}
+              onClick={() => setCategory('l')}
+            >
+              생활 부문
+            </CategoryBox>
+            <CategoryBox
+              active={category === 'g'}
+              onClick={() => setCategory('g')}
+            >
+              게임 부문
+            </CategoryBox>
+          </>
+        ) : (
+          <>
+            <CurrentCategoryBox
+              active={category === 'l'}
+              onClick={() => setCategory('l')}
+            >
+              생활 부문
+            </CurrentCategoryBox>
+            <CurrentCategoryBox
+              active={category === 'g'}
+              onClick={() => setCategory('g')}
+            >
+              게임 부문
+            </CurrentCategoryBox>
+            <CurrentCategoryBox
+              active={category === 'match-l'}
+              onClick={() => setCategory('match-l')}
+            >
+              생활 - 매칭
+            </CurrentCategoryBox>
+            <CurrentCategoryBox
+              active={category === 'match-g'}
+              onClick={() => setCategory('match-g')}
+            >
+              게임 - 매칭
+            </CurrentCategoryBox>
+          </>
+        )}
       </CategoryWrapper>
-      <ListWrapper>{TeamsList}</ListWrapper>
+      <ListWrapper>
+        {category === 'i' || category === 'g' ? TeamsList : MatchList}
+      </ListWrapper>
     </Wrapper>
   );
 };
