+++ src/components/hanseithon/modal/create/create-success/index.ts
@@ -1,0 +1,1 @@
+export { default } from './create-success.component';
