+++ src/components/hanseithon/modal/create/index.ts
@@ -1,0 +1,1 @@
+export { default } from './create-modal.component';
