+++ src/components/hanseithon/modal/join/index.ts
@@ -1,0 +1,1 @@
+export { default } from './join-modal.component';
