+++ src/components/hanseithon/modal/join/join-success/index.ts
@@ -1,0 +1,1 @@
+export { default } from './join-success.component';
