+++ src/components/hanseithon/modal/match/index.ts
@@ -1,0 +1,1 @@
+export { default } from './match-modal.component';
